dev-env-requirements.txt
-r requirements.txt
pytest==3.6.3
pytest-django==3.3.2
flake8==3.5.0
